support for local path firmware flashing

+++ firmware-util.sh
@@ -12,7 +12,10 @@
 #
 
 #where the stuff is
-script_url="https://raw.githubusercontent.com/Dasharo/scripts/add-support-for-custom-image/"
+script_url="https://raw.githubusercontent.com/MrChromebox/scripts/master/"
+# usage: sudo DEV_MODE=true bash firmware-util.sh
+# it will use local version of scripts
+dev_mode=${DEV_MODE:-false}
 
 #ensure output of system tools in en-us for parsing
 export LC_ALL=C
@@ -23,7 +26,7 @@ if grep -q "Chrom" /etc/lsb-release ; then
 	mkdir -p /usr/local/bin
 	cd /usr/local/bin
 else
-	cd /tmp
+	[[ "$dev_mode" == false ]] && cd /tmp
 fi
 
 #check for cmd line param, expired CrOS certs
@@ -33,20 +36,22 @@ else
 	export CURL="curl"
 fi
 
-#get support scripts
-echo -e "\nDownloading supporting files..."
-rm -rf firmware.sh >/dev/null 2>&1
-rm -rf functions.sh >/dev/null 2>&1
-rm -rf sources.sh >/dev/null 2>&1
-$CURL -sLO ${script_url}firmware.sh
-rc0=$?
-$CURL -sLO ${script_url}functions.sh
-rc1=$?
-$CURL -sLO ${script_url}sources.sh
-rc2=$?
-if [[ $rc0 -ne 0 || $rc1 -ne 0 || $rc2 -ne 0 ]]; then
-	echo -e "Error downloading one or more required files; cannot continue"
-	exit 1
+if [[ "$dev_mode" == false ]]; then
+	#get support scripts
+	echo -e "\nDownloading supporting files..."
+	rm -rf firmware.sh >/dev/null 2>&1
+	rm -rf functions.sh >/dev/null 2>&1
+	rm -rf sources.sh >/dev/null 2>&1
+	$CURL -sLO ${script_url}firmware.sh
+	rc0=$?
+	$CURL -sLO ${script_url}functions.sh
+	rc1=$?
+	$CURL -sLO ${script_url}sources.sh
+	rc2=$?
+	if [[ $rc0 -ne 0 || $rc1 -ne 0 || $rc2 -ne 0 ]]; then
+		echo -e "Error downloading one or more required files; cannot continue"
+		exit 1
+	fi
 fi
 
 source ./sources.sh
